requirements.txt
crawl4ai>=0.4.248
google-generativeai==0.3.2
python-dotenv==1.0.1
playwright>=1.49.0
aiohttp>=3.11.0
beautifulsoup4>=4.12.0
lxml>=5.3.0 